+++ src/lib.rs
@@ -40,7 +40,7 @@
 //! # let mut sdmmc_spi = DummySpi;
 //! # let mut sdmmc_cs = DummyCsPin;
 //! # let time_source = DummyTimeSource;
-//! let mut spi_dev = embedded_sdmmc::SdMmcSpi::new(sdmmc_spi, sdmmc_cs);
+//! let mut spi_dev = embedded_sdmmc::SdCard::new(sdmmc_spi, sdmmc_cs);
 //! let block = spi_dev.acquire()?;
 //! println!("Card size {} bytes", block.card_size_bytes()?);
 //! let mut volume_mgr = VolumeManager::new(block, time_source);
@@ -101,7 +101,7 @@ pub use crate::filesystem::{
     Timestamp, MAX_FILE_SIZE,
 };
 pub use crate::sdmmc::Error as SdMmcError;
-pub use crate::sdmmc::{BlockSpi, SdMmcSpi};
+pub use crate::sdmmc::{AcquiredSdCard, SdCard};
 
 mod volume_mgr;
 pub use volume_mgr::VolumeManager;