+++ src/volume_mgr.rs
@@ -237,7 +237,6 @@ where
                 break;
             }
         }
-        drop(dir);
     }
 
     /// Look in a directory for a named file.
@@ -603,7 +602,6 @@ where
                 break;
             }
         }
-        drop(file);
         Ok(())
     }